requirements.txt
grpcio==1.66.1
grpcio-health-checking==1.66.1
grpcio-tools==1.66.1
protobuf==5.28.0
hvac==2.3.0
RPLCD==1.3.1
smbus2==0.4.3
rpi-lgpio==0.6
spidev==3.6
mfrc522-python==0.0.7